requirements.txt
amqp==5.0.6
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dnspython==1.16.0
eventlet==0.31.1
Flask==2.0.1
Flask-MQTT==1.1.1
greenlet==1.1.1
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
MarkupSafe==2.0.1
paho-mqtt==1.5.1
pigpio==1.78
prompt-toolkit==3.0.20
pytz==2021.1
redis==3.5.3
six==1.16.0
smbus==1.1.post2
typing-extensions==3.10.0.0
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.5.0
